==> code/1_data_notes.R <==
p=personal 
o=outdoors
pnc=ultra fine particles
modeled = long term 1 year model at specific location
abs=absorbence - way of measuring concentration of pm2.5


Label	      Details
subjectidp	Subject ID and time point
temp	      Temperature
relhum	    Humidity
gender	
age	
session	Time point (A, B, C)
city	
bmi	
season 	Season at measurement
physical_act	Physical activity
id	Subject ID
date	Date of blood draw
education_adj	Education level
chip	Chip number (for methylation data)
position	Position on the chip (for methylation data)
plate	Plate number (for proteins)
isolation	Technical covariate for transcripts
labeling	Technical covariate for transcripts
hybridization	Technical covariate for transcripts
